requirements.txt
sevenbridges-python